+++ source/extensions/isaacsim.gui.menu/isaacsim/gui/menu/fixme_menu.py
@@ -1,0 +1,95 @@
+import omni.kit.menu.utils
+import omni.ui as ui
+from omni.kit.menu.utils import LayoutSourceSearch, MenuAlignment, MenuItemDescription, MenuLayout
+
+
+class FixmeMenuExtension:
+    class MenuDelegate(ui.MenuDelegate):
+        def get_menu_alignment(self):
+            return MenuAlignment.RIGHT
+
+        # override build fn to not show menu item...
+        def build_item(self, item: ui.MenuHelper):
+            pass
+
+    def __init__(self, ext_id):
+        self._menu_placeholder = [MenuItemDescription(name="FixMe!!!", show_fn=lambda: False)]
+        omni.kit.menu.utils.add_menu_items(self._menu_placeholder, "FixMe", delegate=FixmeMenuExtension.MenuDelegate())
+
+        self.__menu_layout = [
+            MenuLayout.Menu(
+                "FixMe",
+                [
+                    MenuLayout.SubMenu(
+                        "Isaac Utils",
+                        [
+                            # MenuLayout.Item(
+                            #     "Isaac Sim Import Wizard [alpha]", source="Isaac Utils/Isaac Sim Import Wizard [alpha]"
+                            # ),
+                            MenuLayout.Item("Workflows", source="Isaac Utils/Workflows"),
+                            # MenuLayout.Item(
+                            #     "Lula Robot Description Editor", source="Isaac Utils/Lula Robot Description Editor"
+                            # ),
+                            # MenuLayout.Item("Grasp Editor", source="Isaac Utils/Grasp Editor"),
+                            # MenuLayout.Item("Isaac Sim Robot Wizard", source="Isaac Utils/Isaac Sim Robot Wizard"),
+                        ],
+                    ),
+                    MenuLayout.SubMenu(
+                        "Isaac Examples",
+                        [
+                            MenuLayout.Item("Carter URDF", source="Isaac Examples/Import Robots/Carter URDF"),
+                            MenuLayout.Item("Franka URDF", source="Isaac Examples/Import Robots/Franka URDF"),
+                            MenuLayout.Item("Kaya URDF", source="Isaac Examples/Import Robots/Kaya URDF"),
+                            MenuLayout.Item("UR10 URDF", source="Isaac Examples/Import Robots/UR10 URDF"),
+                        ],
+                    ),
+                    MenuLayout.Item("Replicator", source="Replicator"),
+                    MenuLayout.SubMenu(
+                        "Replicator",
+                        [
+                            # have to remove hidden menu items too...
+                            MenuLayout.Item("Capture On Play", source="Replicator/Capture On Play"),
+                            MenuLayout.Item(name="Stop", source="Replicator/Stop"),
+                            MenuLayout.Item(name="Resume", source="Replicator/Resume"),
+                            MenuLayout.Item(name="Pause", source="Replicator/Pause"),
+                            MenuLayout.Item(name="Starting...", source="Replicator/Starting..."),
+                            MenuLayout.Item(name="Starting...", source="Replicator/Starting..."),
+                            MenuLayout.Item(name="Stopping...", source="Replicator/Stopping..."),
+                            MenuLayout.Item(name="Stopping...", source="Replicator/Stopping..."),
+                            MenuLayout.Item(name="Capture On Play", source="Replicator/Capture On Play"),
+                        ],
+                    ),
+                    MenuLayout.SubMenu(
+                        "Profiler",
+                        [
+                            MenuLayout.Item("Start or Stop", source="Profiler/Start\\Stop"),
+                            MenuLayout.Item("Profile Startup (Restart)", source="Profiler/Profile Startup (Restart)"),
+                        ],
+                    ),
+                    MenuLayout.SubMenu(
+                        "Help",
+                        [
+                            MenuLayout.Item("Discover Kit SDK", source="Help/Discover Kit SDK"),
+                            MenuLayout.Item("Developers Manual", source="Help/Developers Manual"),
+                        ],
+                    ),
+                    # MenuLayout.SubMenu(
+                    #     "Physics",
+                    #     [
+                    #         MenuLayout.Item("Physics Authoring Toolbar", source="Physics Toolbar"),
+                    #     ],
+                    # ),
+                    MenuLayout.SubMenu(
+                        "Layout",
+                        [MenuLayout.Item("Quick Save", remove=True), MenuLayout.Item("Quick Load", remove=True)],
+                    ),
+                    MenuLayout.Item("Isaac Sim App Selector", source="Help/Isaac Sim App Selector"),
+                    # MenuLayout.Item("Actions", source="Window/Actions"),
+                ],
+            ),
+        ]
+        omni.kit.menu.utils.add_layout(self.__menu_layout)
+
+    def __del__(self):
+        omni.kit.menu.utils.remove_layout(self.__menu_layout)
+        omni.kit.menu.utils.remove_menu_items(self._menu_placeholder, "FixMe", 9999)