+++ source/extensions/isaacsim.sensors.rtx/python/scripts/menu.py
@@ -286,11 +286,7 @@ class IsaacSensorMenu:
                 sub_menu=rtx_lidar_sub_menu_as_list,
             )
         )
-        self._menu_items = [
-            MenuItemDescription(
-                name="Isaac", glyph="plug.svg", sub_menu=[MenuItemDescription(name="Sensors", sub_menu=menu_items)]
-            )
-        ]
+        self._menu_items = [MenuItemDescription(name="Sensors", glyph="plug.svg", sub_menu=menu_items)]
         add_menu_items(self._menu_items, "Create")
 
     def _get_stage_and_path(self):