+++ source/extensions/isaacsim.sensors.physx/docs/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## [2.1.0] - 2024-10-29
+## [2.2.0] - 2024-10-29
 ### Changed
 - moved examples from menu to browser